refactor: simplify import paths
