+++ apps/client/src/common/hooks/useFadeOutOnInactivity.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 
-import { throttle } from '../../common/utils/throttle';
+import { throttle } from '../utils/throttle';
+
 
 export const useFadeOutOnInactivity = () => {
   const [isMouseMoved, setIsMouseMoved] = useState(false);
